Fix CEF client types and delegate painting
