+++ src/cef/ui/uicefwebview.h
@@ -21,8 +21,8 @@ public:
     virtual ~UICEFWebView();
 
     // CEF-specific methods
-    void onCEFPaint(const void* buffer, int width, int height, const CefRenderHandler::RectList& dirtyRects);
-    void onCEFAcceleratedPaint(const CefAcceleratedPaintInfo& info);
+    void onPaint(const void* buffer, int width, int height, const CefRenderHandler::RectList& dirtyRects);
+    void onAcceleratedPaint(const CefAcceleratedPaintInfo& info);
     void onBrowserCreated(CefRefPtr<CefBrowser> browser);
     
     // Static methods for managing all WebViews
